# docs/swagger.yaml
definitions:
  dto.BlogPost:
    properties:
      body:
        type: string
      id:
        type: integer
      title:
        type: string
      user:
        $ref: '#/definitions/dto.UserInBlogPost'
    type: object
  dto.BlogPostCreateResponse:
    properties:
      body:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  dto.BlogPostPagination:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.BlogPost'
        type: array
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
  dto.BlogPostQueryParams:
    properties:
      current:
        type: integer
      order_by:
        type: string
      page_size:
        type: integer
      q:
        type: string
    type: object
  dto.BlogPostUpdateResponse:
    properties:
      body:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  dto.Comment:
    properties:
      body:
        type: string
      id:
        type: integer
      post:
        $ref: '#/definitions/dto.PostInComment'
      user:
        $ref: '#/definitions/dto.UserInComment'
    type: object
  dto.CommentCreateResponse:
    properties:
      body:
        type: string
      id:
        type: integer
      post_id:
        type: integer
    type: object
  dto.CommentPagination:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.Comment'
        type: array
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
  dto.CommentQueryParams:
    properties:
      current:
        type: integer
      order_by:
        type: string
      page_size:
        type: integer
      q:
        type: string
    type: object
  dto.CommentUpdateResponse:
    properties:
      body:
        type: string
      id:
        type: integer
      post_id:
        type: integer
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 20
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      token:
        type: string
    type: object
  dto.MePasswordRequest:
    properties:
      confirm_password:
        maxLength: 20
        minLength: 8
        type: string
      new_password:
        maxLength: 20
        minLength: 8
        type: string
      old_password:
        maxLength: 20
        minLength: 8
        type: string
    required:
    - confirm_password
    - new_password
    - old_password
    type: object
  dto.MeResponse:
    properties:
      email:
        type: string
      name:
        type: string
    type: object
  dto.MeUpdateRequest:
    properties:
      name:
        type: string
    type: object
  dto.Pagination:
    properties:
      current:
        type: integer
      last_page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  dto.PostInComment:
    properties:
      id:
        type: integer
      title:
        type: string
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 20
        minLength: 3
        type: string
      password:
        maxLength: 20
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  dto.RegisterResponse:
    properties:
      email:
        type: string
      name:
        type: string
    type: object
  dto.User:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  dto.UserInBlogPost:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  dto.UserInComment:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  dto.UserPagination:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.User'
        type: array
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
  dto.UserQueryParams:
    properties:
      current:
        type: integer
      order_by:
        type: string
      page_size:
        type: integer
      q:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      message: {}
    type: object
  response.ValidationError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /blog_posts:
    get:
      consumes:
      - application/json
      description: Get Pagination and Several Posts
      parameters:
      - description: Query Params
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.BlogPostQueryParams'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.BlogPostPagination'
              type: object
      summary: Get Pagination and Several Posts
      tags:
      - blog
    post:
      consumes:
      - application/json
      description: Create a Post
      produces:
      - application/json
      responses:
        "201":
          description: created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.BlogPostCreateResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Create a Post
      tags:
      - blog
  /blog_posts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Post
      produces:
      - application/json
      responses:
        "204":
          description: no content
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete a Post
      tags:
      - blog
    get:
      consumes:
      - application/json
      description: Get detail a post
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.BlogPost'
              type: object
      summary: Get detail a post
      tags:
      - blog
    patch:
      consumes:
      - application/json
      description: Update a Post
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.BlogPostUpdateResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Update a Post
      tags:
      - blog
  /comments:
    get:
      consumes:
      - application/json
      description: Get Pagination and Several Comments
      parameters:
      - description: Query Params
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CommentQueryParams'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CommentPagination'
              type: object
      summary: Get Pagination and Several Comments
      tags:
      - comment
    post:
      consumes:
      - application/json
      description: Create a Comment
      produces:
      - application/json
      responses:
        "201":
          description: created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CommentCreateResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Create a Comment
      tags:
      - comment
  /comments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Comment
      produces:
      - application/json
      responses:
        "204":
          description: no content
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete a Comment
      tags:
      - comment
    get:
      consumes:
      - application/json
      description: Get detail a Comment
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.Comment'
              type: object
      summary: Get detail a Comment
      tags:
      - comment
    patch:
      consumes:
      - application/json
      description: Update a Comment
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CommentUpdateResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Update a Comment
      tags:
      - comment
  /following_blog_posts:
    get:
      consumes:
      - application/json
      description: Get Blog Posts by Following
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.BlogPostPagination'
              type: object
      security:
      - BearerAuth: []
      summary: Get Blog Posts by Following
      tags:
      - blog
  /login:
    post:
      consumes:
      - application/json
      description: Login user and get response JWT Token
      parameters:
      - description: Post
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: bad request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    items:
                      $ref: '#/definitions/response.ValidationError'
                    type: array
                  type: array
              type: object
      summary: Login User
      tags:
      - auth
  /me:
    get:
      consumes:
      - application/json
      description: Get data Logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MeResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get data Logged-in user
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: Update data Logged-in user
      parameters:
      - description: Post
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.MeUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MeResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Update data Logged-in user
      tags:
      - user
  /me/password:
    post:
      consumes:
      - application/json
      description: Get data Logged-in user
      parameters:
      - description: Post
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.MePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get data Logged-in user
      tags:
      - user
  /register:
    post:
      consumes:
      - application/json
      description: register a new user
      parameters:
      - description: Post
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.RegisterResponse'
              type: object
        "400":
          description: bad request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    items:
                      $ref: '#/definitions/response.ValidationError'
                    type: array
                  type: array
              type: object
      summary: Register a new User
      tags:
      - auth
  /users:
    get:
      consumes:
      - application/json
      description: Get Pagination and Several Users
      parameters:
      - description: Query Params
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UserQueryParams'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserPagination'
              type: object
      summary: Get Pagination and Several Users
      tags:
      - user
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Get Detail User
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.User'
              type: object
      summary: Get Detail User
      tags:
      - user
  /users/{id}/follow:
    post:
      consumes:
      - application/json
      description: Following User
      produces:
      - application/json
      responses:
        "202":
          description: accepted
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Following User
      tags:
      - user
  /users/{id}/unfollow:
    post:
      consumes:
      - application/json
      description: UnFollowing User
      produces:
      - application/json
      responses:
        "202":
          description: accepted
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: UnFollowing User
      tags:
      - user
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"